+++ fast_nav/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0
+- Adds context extensions
+- Adds support for Cupertino routes
+- Renames `navigatorName` parameter to `navigator`
+- `NestedNavigator` `name` field is now optional
+
 ## 0.3.0
 - Removes duplicate prevention and simplifies code
 - `NestedNavigator.observers` is now of type `List` instead of `Iterable`